Backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
editdistance==0.8.1
efficientnet==1.0.0
email_validator==2.2.0
essential-generators==1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.2
fastapi-cli==0.0.5
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
gliner==0.2.10
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.7
labelImg==1.8.6
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.19.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pydantic==1.10.9
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
rich==13.8.0
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
termplotlib==0.3.9
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.88
ultralytics-thop==2.0.6
urllib3==1.26.16
uvicorn==0.30.6
uvloop==0.20.0
validators==0.34.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.1
